Cleanedup and moved exceptions into separate file
